+++ hbase/hbase-api/src/main/java/pl/com/sages/hbase/api/loader/LoadMovieTagsData.java
@@ -34,7 +34,7 @@ public class LoadMovieTagsData {
             int userId = Integer.parseInt(data[0]);
             int movieId = Integer.parseInt(data[1]);
             String tag = data[2];
-            long timestamp = Long.parseLong(data[3]);
+            int timestamp = Integer.parseInt(data[3]);
 
             count++;
             if (count % 1000 == 0) {
